[FIX] 그룹 & 목표 일부 수정

+++ src/main/java/com/slide_todo/slide_todoApp/dto/goal/IndividualGoalTodoDTO.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -16,7 +17,7 @@ public class IndividualGoalTodoDTO {
     private Long memberId;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Double progress;
+    private BigDecimal progress;
     private List<IndividualTodoDto> todos;
 
     @Getter
